==> WAC_2025/analysis/2_Evaluations.Rmd <==
---
title: "2_Evaluations"
output: html_document
date: "2025-10-13"
editor_options: 
  chunk_output_type: console
---

```{r include=FALSE}
library(MASS)
library(corrplot)
library(dplyr)
library(tidyr)
library(ggplot2)
library(RColorBrewer)
#install.packages("ggrepel")
library(ggrepel)
#install.packages("irr")
library(irr)
#install.packages("metan")
library(metan)
library(tidyverse)
library(here)
library(lme4)
library(GGally)
```

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = here::here())
```

Make sure data is normal

```{r read}
data <- read_csv("data/Intermediate/data_shared_across_locations.csv")
```

```{r}
# rename all entries for presentable IDs
# alternativly make all plots color coded by base/selection/other

# AVC1 = Maize_base_WI1
# AVC2 = Maize_base_WI2
# AVC3 = Maize_base_WI3
# AVC4 = Maize_base_WI4
# AVC5 = Maize_base_WI5
# AVC6 = WI1_Maize
# AVC7 = WI2_Maize
# AVC9 = WI3_Maize
# AVC11 = WI4_Maize
# AVC14 = WI5_Maize

# IWAF-L-MN1 = MN1_IWG
# IWAF-H-MN1 = MN2_IWG
# IWAF-L-KS1 = KS1_IWG
# IWAF-H-KS2 = KS2_IWG
# IWAF-L-WI1 = WI1_IWG
# IWAF-H-WI1 = WI2_IWG
# IWAF-L-Base = IWG_Base1
# IWAF-H-Base = IWG_Base2

# UMN5560 = MN1_Maize
# UMN5561 = MN2_Maize
# UMN4352 = 
# UMN5553 = 
# UMN4353 = 
# UMN5555 = 
# IAFAL-C3 = Falcata2
# INTALF20 = WI6_Maize
# Larry15 = Falcata1
# NY1713 = 
# NY19-45 = 
```

Box Plots

```{r boxplot}
ggplot(data, aes(x = Entry, y = First_Summer_ALF_Biomass)) +
  geom_boxplot(na.rm = TRUE, fill = "lightblue", color = "black") +
  facet_wrap(~ LOC, ncol = 1) +  # one plot per location stacked vertically
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(title = "Alfalfa biomass by Entry across Locations",
       x = "Entry",
       y = "Biomass (kg/ha)")

```

```{r boxplot2}
colors <- brewer.pal(n = 3, name = "Set2")  # 3 locations

# order entries by selection category
data %>%
  mutate(selected_cat = as.factor(case_when(
    selected == "selected" ~ "B",
    selected == "base" ~ "A",
    TRUE ~ "other"
  ))) %>%
  arrange(selected_cat, Entry) %>%  # ensures grouping by selection
  mutate(Entry = factor(Entry, levels = unique(Entry))) %>%  # lock in order for ggplot
ggplot(aes(x = Entry, y = First_Summer_ALF_Biomass, fill = LOC)) +
  geom_boxplot(na.rm = TRUE, position = position_dodge(width = 0.8)) +
  scale_fill_manual(values = colors) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(title = "Alfalfa biomass by Entry and Location",
       x = "Entry",
       y = "Biomass (kg/ha)",
       fill = "Location")
```

Histograms - Check Normality and Transform

```{r hist}
colnames(data)

hist(data$Emergence_Stand_count) # left skewed

hist(data$Emergence_Alfalfa_Height) # not normal
hist(log(data$Emergence_Alfalfa_Height)) # bimodal

hist(data$First_Summer_Alfalfa_Height) # bimodal
hist(log(data$First_Summer_Alfalfa_Height)) # bimodal

hist(data$First_Summer_IWG_Height) # outliers
hist(log(data$First_Summer_IWG_Height))

hist(data$First_Summer_Maturity)

hist(data$First_Summer_Alfalfa_Stand_count) # left skewed

hist(data$First_Summer_Lodging) # right skewed
hist(log(data$First_Summer_Lodging)) # uniform

hist(data$First_Yr_Regrowth_Vigor)

hist(data$First_Summer_ALF_Biomass) # right skewed 
hist(sqrt(data$First_Summer_ALF_Biomass))

hist(data$First_Summer_IWG_Biomass) # outlier
hist(log(data$First_Summer_IWG_Biomass))

hist(data$First_Summer_Grain_Biomass)

```

```{r}
# is Alfalfa Height driven by lodging 
# split data by high and low lodging and see if those alfalfa hieghts are normal 
low_lodging <- data %>%
  filter(First_Summer_Lodging %in% c(0, 1, 2))

high_lodging <- data %>%
  filter(First_Summer_Lodging %in% c(2, 3, 4, 5, 6, 7, 8, 9))

no_lodging <- data %>%
  filter(First_Summer_Lodging %in% is.na(First_Summer_Lodging))

# WI didn't have lodging values so they are all within:   | is.na(First_Summer_Lodging) 

hist(low_lodging$First_Summer_Alfalfa_Height)
hist(sqrt(low_lodging$First_Summer_Alfalfa_Height))
hist(high_lodging$First_Summer_Alfalfa_Height)
hist(sqrt(high_lodging$First_Summer_Alfalfa_Height))
hist(no_lodging$First_Summer_Alfalfa_Height)

# Alfalfa with a plot lodging score of 0,1,2 have a normally distributed height when transformed 
# Alfalfa with lodging score 3,4,5,6,7,8,9 have a different normally distributed height when transformed
# Alfalfa from WI also have a normal distribution 
```

```{r}
# Are entries different by height ; must account for lodging 
model <- lmer(First_Summer_Alfalfa_Height ~ First_Summer_Lodging + Entry*LOC + (1|LOC:Rep), data=data)
anova(model)

data_model <- data
data_model <- broom.mixed::augment(model, data = data) # add residuals to data file paired with the correct rows
data_model <- data_model %>% rename(First_Summer_ALF_Height = .resid) # rename residual column 
hist(data_model$First_Summer_ALF_Height) # visual check 
# check if that's coming from a normal distribution 
shapiro.test(data_model$First_Summer_ALF_Height)
# W : 1 is normal 
# p-val: Probability that the data are not significantly different from normal

data_model <- data_model %>%
  select(-.fitted, -.hat, -.cooksd, -.fixed, -.mu, -.offset,
         -.sqrtXwt, -.sqrtrwt, -.weights, -.wtres) # remove extra columns 
```

```{r}
# Are entries different by height ; must account for lodging 
model2 <- lmer(First_Summer_IWG_Height ~ First_Summer_Lodging + Entry*LOC + (1|LOC:Rep), data=data)
anova(model2)

data_model2 <- data
data_model2 <- broom.mixed::augment(model2, data = data) # add residuals to data file paired with the correct rows
data_model2 <- data_model2 %>% rename(First_Summer_IWG_height = .resid) # rename residual column 
hist(data_model2$First_Summer_IWG_height) # visual check 
# check if that's coming from a normal distribution 
shapiro.test(data_model2$First_Summer_IWG_height)
# W : 1 is normal 
# p-val: Probability that the data are not significantly different from normal
# sig p-val (<0.05) = no different than expected from normal data
```

```{r norm}
data$First_Summer_IWG_Height_log <- log(data$First_Summer_IWG_Height)
data$First_Summer_ALF_Biomass_sqrt <- sqrt(data$First_Summer_ALF_Biomass)
data$First_Summer_IWG_Biomass_log <- log(data$First_Summer_IWG_Biomass)
data$LOCserp <- paste0(data$LOC,"_",data$serpentine)
data_model$LOCserp <- paste0(data_model$LOC,"_",data_model$serpentine)
data <- merge(data, data_model[,c("LOCserp", "First_Summer_ALF_Height")], by="LOCserp", all.x = TRUE)

# make data frame of normal values
data2 <- data[,c('LOC', 'Entry', 'Rep', 'Row', 'Col', 'Row2', 'Col2', 'serpentine', 'selected', 'selection_location', 'selected_type', 'First_Summer_ALF_Height', 'First_Summer_IWG_Height_log', 'First_Summer_Maturity', 'First_Yr_Regrowth_Vigor', 'First_Summer_ALF_Biomass_sqrt', 'First_Summer_IWG_Biomass_log', 'First_Summer_Lodging')]
```

Correlation 

```{r corr}
# Since there are both continuous and ordinal traits, use spearman instead of pearson

# make better label names
data2$Alfalfa_Height <- data2$First_Summer_ALF_Height
data2$IWG_Biomass <- data2$First_Summer_IWG_Biomass_log
data2$IWG_Height <- data2$First_Summer_IWG_Height_log
data2$Alfalfa_Maturity <- data2$First_Summer_Maturity
data2$Alfalfa_Regrowth <- data2$First_Yr_Regrowth_Vigor
data2$Alfalfa_Biomass <- data2$First_Summer_ALF_Biomass_sqrt
data2$Lodging <- data2$First_Summer_Lodging


correlations = cor(data2[,c(13:ncol(data2))], use = "pairwise.complete.obs", method = "spearman")
corrplot.mixed(correlations, order = 'AOE')
## leave blank on non-significant coefficient
## add significant correlation coefficients
testRes = cor.mtest(data2[,c(13:ncol(data2))], conf.level = 0.95, method = "spearman")
corrplot(correlations, p.mat = testRes$p, method = 'circle', type = 'lower', insig='blank',
         addCoef.col ='black', number.cex = 0.8, order = 'AOE', diag=FALSE)

# Just Alfalfa Traits
correlations2 = cor(data2[,c("Alfalfa_Height", "Alfalfa_Maturity", "Alfalfa_Regrowth", "Alfalfa_Biomass")], use = "pairwise.complete.obs", method = "spearman")
#corrplot.mixed(correlations2, order = 'AOE')
## leave blank on non-significant coefficient
## add significant correlation coefficients
testRes = cor.mtest(data2[,c("Alfalfa_Height", "Alfalfa_Maturity", "Alfalfa_Regrowth", "Alfalfa_Biomass")], conf.level = 0.95, method="spearman")
corrplot(correlations2, p.mat = testRes$p, method = 'circle', type = 'lower', insig='blank',
         addCoef.col ='black', number.cex = 0.8, order = 'AOE', diag=FALSE)

# not in spearman but a cool plot
data2 %>%
  ggpairs(columns = c("Alfalfa_Height", "Alfalfa_Maturity", "Alfalfa_Regrowth", "Alfalfa_Biomass"),
          aes(color = selection_location))+
  theme(strip.text = element_text(size = 18))
```

Stability or consistency of performance across environments

```{r ranking}
# average entry value per location
entry_means <- data2 %>%
  group_by(LOC, Entry) %>%
  summarise(mean_biomass = mean(First_Summer_ALF_Biomass_sqrt, na.rm = TRUE)) %>%
  ungroup()

# rank entries per location
ranked <- entry_means %>%
  group_by(LOC) %>%
  mutate(rank = rank(-mean_biomass, ties.method = "average")) %>%
  ungroup()

rank_wide <- ranked %>%
  select(LOC, Entry, rank) %>%
  pivot_wider(names_from = LOC, values_from = rank)
```

```{r stability}
# Overall stability (mean and sd)
rank_summary <- rank_wide %>%
  mutate(mean_rank = rowMeans(across(c(KS,NY,WI)), na.rm = TRUE),
         sd_rank = apply(across(c(KS,NY,WI)), 1, sd, na.rm = TRUE)) %>%
  arrange(mean_rank)
rank_summary
```

```{r stabilityplot}
ggplot(rank_summary, aes(x = mean_rank, y = sd_rank, label = Entry)) +
geom_point() +
geom_text_repel() +
labs(x = "Mean Rank (Overall Performance)",
y = "SD of Rank (Stability)",
title = "Genotype Stability Across Locations")

```

```{r rankcorr}
# pairwise rank correlation of biomass
cor(rank_wide$WI, rank_wide$NY, method = "spearman", use = "pairwise.complete.obs")
cor(rank_wide$WI, rank_wide$KS, method = "spearman", use = "pairwise.complete.obs")
cor(rank_wide$NY, rank_wide$KS, method = "spearman", use = "pairwise.complete.obs")
# WI:NY -0.2345865
# WI:KS 0.362406
# NY:KS -0.09401709
```

```{r kendallW}
# Select rank columns
rank_matrix <- rank_wide %>%
  select(WI, NY, KS)

# Kendall’s W
kendall(rank_matrix)
# W = relative agreement (0-1)
# pval = is observed agreement significantly greater than expected by chance
```

Visualize GxE

```{r GGEbiplot}
# GGE biplot to visualize genotype × environment (G×E) interaction and identify which genotypes perform best in which environments
means_wide <- entry_means %>%
  select(LOC, Entry, mean_biomass) %>%
  pivot_wider(names_from = LOC, values_from = mean_biomass)

gge_model <- gge(entry_means, env = LOC, gen = Entry, resp = mean_biomass)
plot(gge_model, type = 1, size.lab.gen = 3,   # smaller label size
     max.overlaps = 15000, # allow more overlaps
     repel = TRUE)   # which-won-where
```

```{r rankingplot}
# visual of rankings
ggplot(ranked, aes(x = LOC, y = rank, group = Entry, color = Entry)) +
  geom_line() +
  geom_point() +
  scale_y_reverse() +  # rank 1 at top
  theme_minimal() +
  labs(title = " ",
       y = "Rank (1 = highest biomass)")
```

```{r save}
#write.csv(data2, 'data/Intermediate/data_shared_across_locations_transformed.csv')
```
