==> WAC_2025/analysis/3_Linear_Models.Rmd <==
---
title: "3_Linear_Models"
output: html_document
date: "2025-10-13"
editor_options: 
  chunk_output_type: console
---

```{r include=FALSE}
library(tidyverse)
#install.packages("here")
library(here)
library(multcompView)
library(lmerTest)
library(lme4)
library(lmtest)
library("PerformanceAnalytics")
library(dplyr)
library(broom.mixed)
library(emmeans)
library(ordinal)
library(kableExtra)
library(tibble)
library(flextable)
library(RVAideMemoire)
library(multcomp)
```

```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = here::here())
```

```{r}
data <- read.csv('data/Intermediate/data_shared_across_locations_transformed.csv')
```

```{r}
#data2 <- data %>%
#  mutate(selected_cat = as.factor(case_when(
#    selected == "selected" ~ "selected",
#    selected == "base" ~ "base",
#    TRUE ~ "other"
#  )))%>%
#  mutate(selected_cat = as.factor(case_when(
#   selection_location == "WI" ~ "selectWI",
#    selection_location == "KS" ~ "select",
#    selection_location == "MN" ~ "select",
#    selected == "base" ~ "base",
#    TRUE ~ "other"
#  )))
```

```{r}
data$LOC <- as.factor(data$LOC)
data$Entry <- as.factor(data$Entry)
data$Rep <- as.factor(data$Rep)
data$selected <- as.factor(data$selected)
data$selection_location <- as.factor(data$selection_location)
data$selected_type <- as.factor(data$selected_type)
data$selected_category <- as.factor(data$selected_category)
```

## Testing significance of Entry on first year traits 

```{r}
# are entries significantly different in first year summer biomass
model <- lmer(First_Summer_ALF_Biomass_sqrt ~ Entry*LOC + (1|LOC:Rep), data=data)
aov.table <- anova(model)
#aov.table
#summary(model)
# Location is significant covariant Pr(>F) = 0.04401
# When controlling for location, Entry is not significant, although AVC3 had significantly lower biomass than other Entries
table = flextable(data = aov.table %>%
                    rownames_to_column(" ")) 

bold(table, bold = TRUE, part = "header")
# shorter table 
table = flextable(data = aov.table %>%
                    select(NumDF, `Pr(>F)`) %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

### what specific entries are doing well in locations
# pairwise comparison 
# Which entries perform significantly differently overall (across all locations)
emm <- emmeans(model, list(pairwise ~ Entry), adjust = "tukey")
pairs_df <- as.data.frame(emm[["pairwise differences of Entry"]]) # Extract pairwise results table
# Keep only significant comparisons (p < 0.05)
sig_pairs <- pairs_df %>%
  filter(p.value < 0.05) %>%
  arrange(p.value)
# Print nicely formatted table
kbl(sig_pairs, digits = 5, caption = "Significant pairwise comparisons of Entry (Tukey-adjusted) <br> biomass ~ Entry*Location + (1|Location)") %>%
  kable_classic(full_width = F, html_font = "Helvetica")

# Which entries differ significantly within each location?
emm_loc <- emmeans(model, list(pairwise ~ Entry|LOC), adjust = "tukey")
pairs_loc <- as.data.frame(emm_loc[["pairwise differences of Entry | LOC"]]) # Extract pairwise results table
# Keep only significant comparisons (p < 0.05)
sig_pairs_loc <- pairs_loc %>%
  filter(p.value < 0.05) %>%
  arrange(p.value)
# Print nicely formatted table
kbl(sig_pairs_loc, digits = 5, caption = "Significant pairwise comparisons of GxE (Tukey-adjusted) <br> biomass ~ Entry*Location + (1|Location)") %>%
  kable_classic(full_width = F, html_font = "Helvetica")
```

```{r}
# are entries significantly different in first summer height 
model2 <- lmer(First_Summer_ALF_Height ~ Entry*LOC + (1|LOC:Rep), data=data)
anova(model2)
```

```{r}
# are entries significantly different in first year vigor
# cumulative link mixed model (CLMM) for ordered but not continuous data (Ordinal categorical response)
data$Alfalfa_Regrowth <- ordered(data$Alfalfa_Regrowth, levels = c(0,1,2,3,4,5)) # set as ordinal and define categories 
data$Alfalfa_Regrowth <- ordered(data$First_Yr_Regrowth_Vigor2, levels = c(0,1,2,3,4,5)) # set as ordinal and define categories 
model3 <- clmm(Alfalfa_Regrowth ~ Entry + (1|LOC:Rep), data=data)
aov_clmm_talbe <- Anova.clmm(model3, type = "II")

table = flextable(data = aov_clmm_talbe %>%
      mutate(`Pr(>Chisq)` = formatC(`Pr(>Chisq)`, format = "e", digits = 2)) %>% 
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

# Which entries differ significantly within each location?
emm_loc <- emmeans(model3, list(pairwise ~ Entry), adjust = "tukey")
pairs_loc <- as.data.frame(emm_loc[["pairwise differences of Entry"]]) # Extract pairwise results table
# Keep only significant comparisons (p < 0.05)
sig_pairs_loc <- pairs_loc %>%
  filter(p.value < 0.05) %>%
  arrange(p.value)
# Print nicely formatted table
kbl(sig_pairs_loc, digits = 5, caption = "Significant pairwise comparisons (Tukey-adjusted)") %>%
  kable_classic(full_width = F, html_font = "Helvetica")

## Although the Entry × Location interaction was significant, sparse representation of entries across locations prevented reliable estimation of entry-specific marginal means

data %>%
  filter(LOC %in% c("NY", "WI")) %>%
  droplevels() %>%
  ggplot(aes(x = Entry, y = Alfalfa_Regrowth))+
  geom_dotplot(
    binaxis = "y",       # stack along the y-axis
    stackdir = "center", # center dots horizontally
    dotsize = 1,       # adjust for density
    fill = "purple",  # color of dots
    color = "black"      # dot outline
  ) +
  facet_wrap(~ LOC, ncol = 1) +  # one plot per location stacked vertically
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(title = "Alfalfa Regrowth by Entry across Locations",
       x = "Entry",
       y = "Regrowth Vigor Rating")
# save output/regrowth_entry_yr1S_dotplot
# output/regrowth_entry_yr1S_dotplot3 1200x815 (with color and size changes lightblue > purple, dotsize 0.6 > 1)

data %>%
  filter(LOC %in% c("NY", "WI")) %>%
  droplevels() %>%
  filter(!str_detect(Entry, "^(UMN|NY)"))%>%
  ggplot(aes(x = Entry, y = Alfalfa_Regrowth, fill = LOC))+
  geom_dotplot(
    binaxis = "y",       # stack along the y-axis
    stackdir = "center", # center dots horizontally
    dotsize = 1.5,       # adjust for density
    color = "black"      # dot outline
  ) +
  scale_fill_manual(
    values = c(
      NY = "purple",  # blue
      WI = "#33a02c"   # green
    )
  ) +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(title = "Alfalfa Regrowth by Entry across Locations",
       x = "Entry",
       y = "Regrowth Vigor Rating")
# save output/regrowth_entry_yr1S_dotplot2

data2 <- data%>%
  filter(LOC %in% c("NY", "WI")) %>%
  droplevels() %>%
  filter(!str_detect(Entry, "^(UMN|NY)"))
model3 <- clmm(Alfalfa_Regrowth ~ Entry*LOC + (1|LOC:Rep), data=data2)
aov_clmm_talbe <- Anova.clmm(model3, type = "II")
table = flextable(data = aov_clmm_talbe %>%
      mutate(`Pr(>Chisq)` = formatC(`Pr(>Chisq)`, format = "e", digits = 2)) %>% 
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")
# Which entries differ significantly within each location?
emm_loc <- emmeans(model3, list(pairwise ~ Entry|LOC), adjust = "tukey")
# Hessian is numerically singular: parameters are not uniquely determined
# variance–covariance matrix of fixed effects cannot be computed

model_NY <- clmm(
  Alfalfa_Regrowth ~ Entry + (1 | Rep),
  data = filter(data2, LOC == "NY")
)
emm_loc <- emmeans(model_NY, list(pairwise ~ Entry), adjust = "tukey")
pairs_loc <- as.data.frame(emm_loc[["pairwise differences of Entry"]]) # Extract pairwise results table
sig_pairs_loc <- pairs_loc %>%
  filter(p.value < 0.05) %>%
  arrange(p.value)
kbl(sig_pairs_loc, digits = 5, caption = "Significant pairwise comparisons (Tukey-adjusted)") %>%
  kable_classic(full_width = F, html_font = "Helvetica")

model_WI <- clmm(
  Alfalfa_Regrowth ~ Entry + (1 | Rep),
  data = filter(data2, LOC == "WI")
)
emm_loc <- emmeans(model_WI, list(pairwise ~ Entry), adjust = "tukey")
# Hessian is numerically singular: parameters are not uniquely determined
# variance–covariance matrix of fixed effects cannot be computed
#pairs_loc <- as.data.frame(emm_loc[["pairwise differences of Entry"]]) # Extract pairwise results table
#sig_pairs_loc <- pairs_loc %>%
#  filter(p.value < 0.05) %>%
#  arrange(p.value)
#kbl(sig_pairs_loc, digits = 5, caption = "Significant pairwise comparisons (Tukey-adjusted)") %>%
#  kable_classic(full_width = F, html_font = "Helvetica")

custom_letters <- c("a", "b", "c", "d", "e", "f", "g","h","i","j","k","l","m","n","o","p","q","r","s","t","u","v","w","x","y","z")
emm_ny <- emmeans(model_NY, list(pairwise ~ Entry), adjust = "tukey")
cld_ny <- cld(emm_ny, Letters=custom_letters) # Create compact letter display
rownames(cld_ny) <- NULL
cld_ny<- cld_ny %>%
  select(Entry, .group)
cld_ny  |> 
  data.frame() |> 
  kbl(align = "c") |> 
  kable_paper(full_width = F) |> 
  column_spec(1, bold = T)

data2 %>%
  filter(LOC == "NY")%>%
  group_by(Entry) %>%
  mutate(Alfalfa_Regrowth2 = as.numeric(Alfalfa_Regrowth))%>%
  summarise(
    count = n(),
    mean = mean(Alfalfa_Regrowth2,na.rm=TRUE),
    stddev = sd(Alfalfa_Regrowth2, na.rm=TRUE),
    meansd_l = mean - stddev,
    meansd_u = mean + stddev
  )%>%
  left_join(cld_ny, by = "Entry") %>%
  arrange(desc(mean))%>%
  ggplot(aes(x = reorder(Entry, mean), y=mean)) + 
        geom_point()+
  geom_errorbar(aes(ymin = meansd_l, ymax = meansd_u), width=0.1)+
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(title = "Alfalfa Regrowth NY",
       x = "Entry",
       y = "Regrowth Vigor Rating")+
  geom_text(aes(x = Entry, label = .group),
            color = "black", vjust = 0, hjust = 0)
# save output/regrowth_yr1S
```

```{r}
# are entries significantly different in first year maturity
# cumulative link mixed model (CLMM) for ordered but not continuous data (Ordinal categorical response)
data$Alfalfa_Maturity <- ordered(data$Alfalfa_Maturity, levels = c(0,1,2,3,4,5,6,7,8,9)) # set as ordinal and define categories 
data$Alfalfa_Maturity <- ordered(data$First_Summer_Maturity, levels = c(0,1,2,3,4,5,6,7,8,9)) # set as ordinal and define categories 
model4 <- clmm(Alfalfa_Maturity ~ Entry*LOC + (1|LOC:Rep), data=data)
summary(model4)
aov_clmm_table <- Anova.clmm(model4, type = "II")
#aov_clmm_table

table = flextable(data = aov_clmm_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")


# Which entries differ significantly within each location?
emm_loc <- emmeans(model4, list(pairwise ~ Entry|LOC), adjust = "tukey")
pairs_loc <- as.data.frame(emm_loc[["pairwise differences of Entry | LOC"]]) # Extract pairwise results table
# Keep only significant comparisons (p < 0.05)
sig_pairs_loc <- pairs_loc %>%
  filter(p.value < 0.05) %>%
  arrange(p.value)
# Print nicely formatted table
kbl(sig_pairs_loc, digits = 5, caption = "Significant pairwise comparisons (Tukey-adjusted)") %>%
  kable_classic(full_width = F, html_font = "Helvetica")
```



## Do selected alfalfa differ from base alfalfa?

# test whether selection has led to significant phenotypic differences across traits

Biomass

selected : base, other, selected

```{r}
# do entries selected for intercropping differ than unselected entries
model5 <- lmer(First_Summer_ALF_Biomass_sqrt ~ selected*LOC + (1|LOC:Rep), data=data)
anova(model5)
```

```{r}
# tests if selection effects are consistent or location-dependent
model7 <- lmer(First_Summer_ALF_Biomass_sqrt ~ selected*LOC + (1|Entry) + (1|LOC:Rep), data = data)
anova(model7)
emmeans(model7, pairwise ~ selected | LOC)
```

test whether selection location and neighbor identity led to significant differences 

selection_cat : 
base_IWG 
KS_IWG, MN_IWG, WI_IWG
base_Maize
WI_Maize
MN_Maize
other  

```{r}
# Is the selection location and neighbor significant to entry performance

data <- data %>%
  mutate(sel_base = if_else(
    Entry %in% c("IWAF-L-Base", "IWAF-H-Base"),
    "base_IWG",
    selected
  ))%>%
  mutate(sel_base2 = if_else(
    sel_base %in% c("base"), "base_Maize", sel_base
  ))%>%
  mutate(selection_cat = if_else(
    is.na(selected_type) | selected_type == "",
    sel_base2,
    selected_type
  ))

data$selection_cat <- as.factor(data$selection_cat)

model8 <- lmer(First_Summer_ALF_Biomass_sqrt ~ selection_cat*LOC + (1|Entry) + (1|LOC:Rep), data=data)
aov_select_table <- anova(model8)
table = flextable(data = aov_select_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

table = flextable(data = aov_select_table %>%
                    select(NumDF, `Pr(>F)`) %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

# Which entries differ significantly within each location?
emm_loc <- emmeans(model8, list(pairwise ~ selected_type|LOC), adjust = "tukey")
pairs_loc <- as.data.frame(emm_loc[["pairwise differences of selected_type | LOC"]]) # Extract pairwise results table
# Keep only significant comparisons (p < 0.05)
sig_pairs_loc <- pairs_loc %>%
  filter(p.value < 0.05) %>%
  arrange(p.value)
# shorten
sig_pairs_loc <- pairs_loc %>%
  filter(p.value < 0.05) %>%
  arrange(p.value)%>%
  select(`2`, LOC, p.value)%>%
  rename(" " = `2`)
# Print nicely formatted table
kbl(sig_pairs_loc, digits = 5, caption = "Significant pairwise comparisons (Tukey-adjusted)") %>%
  kable_classic(full_width = F, html_font = "Helvetica")
```

selected_category:
base_H, base_L
base_WI 
KS_IWG, MN_IWG, WI_IWG
MN_Maize  
WI_Maize

```{r}
model6 <- lmer(First_Summer_ALF_Biomass_sqrt ~ selected_category*LOC + (1|Entry) + (1|LOC:Rep), data=data)
aov_select_table <- anova(model6)
table = flextable(data = aov_select_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

table = flextable(data = aov_select_table %>%
                    select(NumDF, `Pr(>F)`) %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")
```

```{r}
model10 <- lmer(First_Summer_ALF_Biomass_sqrt ~ selected_type*LOC + (1|Entry) + (1|LOC:Rep), data=data)
anova(model10)
```

```{r}
AIC(model5, model6, model7, model8, model10)
# not fitted to same number of observations
# model 10 : KS_IWG, MN_IWG, WI_IWG, MN_Maize, WI_Maize -- selected_type*LOC + (1|Entry) + (1|LOC:Rep)
```

```{r}
emm_loc <- emmeans(model10, list(pairwise ~ selected_type|LOC), adjust = "tukey")
pairs_loc <- as.data.frame(emm_loc[["pairwise differences of selected_type | LOC"]]) # Extract pairwise results table
# Keep only significant comparisons (p < 0.05)
sig_pairs_loc <- pairs_loc %>%
  filter(p.value < 0.06) %>%
  arrange(p.value)
# shorten
sig_pairs_loc <- pairs_loc %>%
  filter(p.value < 0.06) %>%
  arrange(p.value)%>%
  select(`2`, LOC, p.value)%>%
  rename(" " = `2`)
# Print nicely formatted table
kbl(sig_pairs_loc, digits = 5, caption = "Significant pairwise comparisons (Tukey-adjusted)") %>%
  kable_classic(full_width = F, html_font = "Helvetica")
# save output/biomass_selcat_yr1S_emmean

data %>%
  filter(LOC == "NY", selected_type == "MN_IWG") %>%
  summarise(mean_value = mean(First_Summer_ALF_Biomass_sqrt, na.rm = TRUE))
data %>%
  filter(LOC == "NY", selected_type == "WI_Maize") %>%
  summarise(mean_value = mean(First_Summer_ALF_Biomass_sqrt, na.rm = TRUE))
data %>%
  filter(LOC == "NY", selected_type == "MN_Maize") %>%
  summarise(mean_value = mean(First_Summer_ALF_Biomass_sqrt, na.rm = TRUE))
```

Plant Height

tests if selection effects are consistent or location-dependent

```{r}
# do entries selected for intercropping differ than unselected entries
model9 <- lmer(First_Summer_ALF_Height ~ selected*LOC + (1|LOC:Rep), data=data)
anova(model9)
```

```{r}
# tests if selection effects are consistent or location-dependent
model11 <- lmer(First_Summer_ALF_Height ~ selected*LOC + (1|Entry) + (1|LOC:Rep), data = data)
anova(model11)
#emmeans(model7, pairwise ~ selected | LOC)
```

```{r}
# Is the selection location and neighbor significant to entry performance
model12 <- lmer(First_Summer_ALF_Height ~ selected_type*LOC + (1|Entry) + (1|LOC:Rep), data=data)
anova(model12)
```

```{r}
model13 <- lmer(First_Summer_ALF_Height ~ selection_cat*LOC + (1|Entry) + (1|LOC:Rep), data=data)
```

```{r}
model14 <- lmer(First_Summer_ALF_Height ~ selected_category*LOC + (1|Entry) + (1|LOC:Rep), data=data)
```

```{r}
AIC(model9, model11, model12, model13, model14)
# not fitted to same number of observations
# model 12 : KS_IWG, MN_IWG, WI_IWG, MN_Maize, WI_Maize -- selected_type*LOC + (1|Entry) + (1|LOC:Rep)
```

Summer Maturity 

tests if selection effects are consistent or location-dependent

Could be interesting if selection for delayed maturity (matching delayed summer harvest) has occurred

```{r}
model15 <- clmm(Alfalfa_Maturity ~ selected_type * LOC + (1|LOC:Rep) + (1|Entry), data = data)
summary(model15)
aov_clmm_talbe <- Anova.clmm(model15, type = "II")

table = flextable(data = aov_clmm_talbe %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")


```

Regrowth Vigor

```{r}
model16 <- clmm(Alfalfa_Regrowth ~ selected_type * LOC + (1|Entry) + (1|LOC:Rep), data = data)
summary(model16)
aov_clmm_table <- Anova.clmm(model16, type = "III")

table = flextable(data = aov_clmm_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

```


# selection category might only have two entries within a category
# what about more general categories?

# neighbor selection 

```{r}
data <- data %>%
  mutate(sel_cat2 = case_when(
    selection_cat == "other" ~ "other",
    selection_cat %in% c("WI_IWG", "MN_IWG", "KS_IWG") ~ "IWG",
    selection_cat %in% c("WI_Maize", "MN_Maize") ~ "Maize",
    selection_cat == "base_Maize" ~ "base_Maize",
    selection_cat == "base_IWG" ~ "base_IWG"
  ))

data$sel_cat2 <- as.factor(data$sel_cat2)
unique(data$sel_cat2)

model17 <- lmer(First_Summer_ALF_Biomass_sqrt ~ sel_cat2*LOC + (1|Entry) + (1|LOC:Rep), data=data)
aov_select_table <- anova(model17)
table = flextable(data = aov_select_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

table = flextable(data = aov_select_table %>%
                    dplyr::select(NumDF, `Pr(>F)`) %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

model17 <- clmm(Alfalfa_Regrowth ~ sel_cat2 * LOC + (1|Entry) + (1|LOC:Rep), data = data)
summary(model17)
aov_clmm_table <- Anova.clmm(model17, type = "III")

table = flextable(data = aov_clmm_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")


```

# location selected in 

```{r}
data <- data %>%
  mutate(sel_cat3 = case_when(
    selection_cat == "other" ~ "other",
    selection_cat %in% c("WI_IWG", "WI_Maize") ~ "WI",
    selection_cat %in% c("MN_IWG", "MN_Maize") ~ "MN",
    selection_cat == "KS_IWG" ~ "KS",
    selection_cat == "base_Maize" ~ "base_Maize",
    selection_cat == "base_IWG" ~ "base_IWG"
  ))

data$sel_cat3 <- as.factor(data$sel_cat3)
unique(data$sel_cat3)

model18 <- lmer(First_Summer_ALF_Biomass_sqrt ~ sel_cat3*LOC + (1|Entry) + (1|LOC:Rep), data=data)
aov_select_table <- anova(model18)
table = flextable(data = aov_select_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

table = flextable(data = aov_select_table %>%
                    dplyr::select(NumDF, `Pr(>F)`) %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")


model19 <- clmm(Alfalfa_Regrowth ~ sel_cat3 * LOC + (1|Entry) + (1|LOC:Rep), data = data)
summary(model19)
aov_clmm_table <- Anova.clmm(model19, type = "III")

table = flextable(data = aov_clmm_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")
```

```{r}
model20 <- lmer(First_Summer_ALF_Biomass_sqrt ~ sel_cat2*LOC + sel_cat3*LOC + (1|Entry) + (1|LOC:Rep), data=data)
aov_select_table <- anova(model20)
table = flextable(data = aov_select_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")

table = flextable(data = aov_select_table %>%
                    dplyr::select(NumDF, `Pr(>F)`) %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")


model21 <- clmm(Alfalfa_Regrowth ~ sel_cat3*LOC + sel_cat2*LOC + (1|Entry) + (1|LOC:Rep), data = data)
summary(model21)
aov_clmm_table <- Anova.clmm(model19, type = "III")

table = flextable(data = aov_clmm_table %>%
                    rownames_to_column(" ")) 
bold(table, bold = TRUE, part = "header")
```

## Likelihood ratio test

```{r}
# determine whether being selected for intercropping is statistically significant
# is their a significant difference between the two models in the liklihood of observing given data given normal distribution 
model_Ha <- lmer(First_Summer_ALF_Biomass_sqrt ~ selection_cat*LOC + (1 | Entry) + (1|LOC:Rep), data=data, REML=FALSE) # alternate hypothesis : intercropping selection influences biomass

model_H0 <- lmer(First_Summer_ALF_Biomass_sqrt ~ 1 + (1 | Entry) + (1|LOC:Rep), data=data , REML=FALSE) # null hypothesis : selection doesn't matter to biomass

anova(model_H0, model_Ha)
# Pr(>Chisq) = 0.08768
# Alternative model with selection does not significantly improve model fit
# Being selected for intercropping does not significantly affect First_Summer_ALF_Biomass (after accounting for Entry, Location, and Rep)

LRT <- lrtest(model_H0, model_Ha)
LRT
# 1 & 2 : Null and Alternative model
# Df : df or n parameters
# Loglik : log-liklihood values
# Chisq : likelihood ratio test statistic (difference in log-likelihood between models)
# Pr(>Chisq): p-value associated with likelihood ratio test
## p-value < 0.05 : reject null hypothesis. Alternative model better explains data

emmeans(model_Ha, pairwise ~ selection_cat)
plot(emmeans(model_Ha, ~ selection_cat))
anova(model_Ha)
```

```{r}
# determine whether being selected for intercropping in WI is statistically significant
# is their a significant difference between the two models in the liklihood of observing given data given normal distribution 
model_Ha <- lmer(First_Summer_ALF_Biomass_sqrt ~ selected_cat2 + (1 | Entry) + (1|LOC:Rep), data=data2) # alternate hypothesis : intercropping selection influences biomass

model_H0 <- lmer(First_Summer_ALF_Biomass_sqrt ~ 1 + (1 | Entry) + (1|LOC:Rep), data=data2) # null hypothesis : selection doesn't matter to biomass

LRT <- lrtest(model_H0, model_Ha)
LRT
# 1 & 2 : Null and Alternative model
# Df : df or n parameters
# Loglik : log-liklihood values
# Chisq : likelihood ratio test statistic (difference in log-likelihood between models)
# Pr(>Chisq): p-value associated with likelihood ratio test
## p-value < 0.05 : reject null hypothesis. Alternative model better explains data

emmeans(model_Ha, pairwise ~ selected_cat2)
plot(emmeans(model_Ha, ~ selected_cat2))
```

```{r}
# Are WI populations selected from a specific WI base population different between the base and selected
# equal variance t-test
## base > selected
## AVC1 > AVC6
## AVC2 > AVC7
## AVC3 > AVC9
## AVC4 > AVC11
## AVC5 > AVC 14
## IWAF-L-Base > IWAF-L-WI1
## IWAF-H-Base > IWAF-H-WI1

# Manually 
# t = (x_bar1 - x_bar2)/ sqrt(sd^2(1/n1 + 1/n2))
x_bar1 <- data %>%
  filter(Entry == "AVC1") %>%
  summarise(mean_value = mean(First_Summer_ALF_Biomass_sqrt, na.rm = TRUE)) %>%
  pull(mean_value)
x_bar2 <- data %>%
  filter(Entry == "AVC6") %>%
  summarise(mean_value = mean(First_Summer_ALF_Biomass_sqrt, na.rm = TRUE)) %>%
  pull(mean_value)
sd <- data %>%
  filter(Entry %in% c("AVC1", "AVC6")) %>%
  summarise(sd_value = sd(First_Summer_ALF_Biomass_sqrt, na.rm = TRUE)) %>%
  pull(sd_value)
n1 <- data %>%
  filter(Entry == "AVC1", !is.na(First_Summer_ALF_Biomass_sqrt)) %>%
  summarise(n = n()) %>%
  pull(n)
n2 <- data %>%
  filter(Entry == "AVC6", !is.na(First_Summer_ALF_Biomass_sqrt)) %>%
  summarise(n = n()) %>%
  pull(n)
df <- n1 + n2 - 2
t = (x_bar1 - x_bar2)/ sqrt(sd^2 * (1/n1 + 1/n2))
# critical t-value for one-tailed t-test with a=0.05 and df=21 : 1.721
# critical t-value for two-tailed t-test with a=0.05 and df=21 : 2.080

# t-test function 
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("AVC1", "AVC6")),
       var.equal = TRUE)
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("AVC2", "AVC7")),
       var.equal = TRUE)
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("AVC3", "AVC9")),
       var.equal = TRUE)
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("AVC4", "AVC11")),
       var.equal = TRUE)
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("AVC5", "AVC14")),
       var.equal = TRUE)
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("IWAF-L-Base", "IWAF-L-WI1")),
       var.equal = TRUE)
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("IWAF-H-Base", "IWAF-H-WI1")),
       var.equal = TRUE)
```

```{r}
# Are any populations selected from the base population different between the base and selected
# equal variance t-test
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("IWAF-L-Base", "IWAF-L-KS1")),
       var.equal = TRUE)
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("IWAF-H-Base", "IWAF-H-KS2")), # only exists in KS
       var.equal = TRUE)
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("IWAF-L-Base", "IWAF-L-MN1")),
       var.equal = TRUE)
t.test(First_Summer_ALF_Biomass_sqrt ~ Entry,
       data = filter(data, Entry %in% c("IWAF-H-Base", "IWAF-H-MN1")),
       var.equal = TRUE)
```
